+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 These are the changes that have not been released yet. The release notes have not yet been written.
 
+**FEATURE ENCHANCEMENTS:**
+
+- Add idempotency flag to `build`, `load`, and `unload` commands.
+
 ## v0.3.5 (2022-10-18)
 
 **FEATURE ENCHANCEMENTS:**